Yesod.Auth

Subsite

data Auth

type AuthRoute = Route Auth

data family Route a

The type-safe URLs associated with a site argument.

data AuthPlugin m

Constructors

AuthPlugin

Fields

apName :: Text

apDispatch :: Method -> [Piece] -> GHandler Auth m ()

apLogin :: forall s. (Route Auth -> Route m) -> GWidget s m ()

getAuth :: a -> Auth

class (Yesod m, PathPiece (AuthId m), RenderMessage m FormMessage) => YesodAuth m where

Associated Types

type AuthId m

Methods

loginDest :: m -> Route m

Default destination on successful login, if no other destination exists.

logoutDest :: m -> Route m

Default destination on successful logout, if no other destination exists.

getAuthId :: Creds m -> GHandler s m (Maybe (AuthId m))

Determine the ID associated with the set of credentials.

authPlugins :: m -> [AuthPlugin m]

Which authentication backends to use.

loginHandler :: GHandler Auth m RepHtml

What to show on the login page.

renderAuthMessage

Arguments

:: m

-> [Text]
languages

-> AuthMessage

-> Text

Used for i18n of messages provided by this package.

redirectToReferer :: m -> Bool

After login and logout, redirect to the referring page, instead of loginDest and logoutDest. Default is False.

authHttpManager :: m -> Manager

onLogin :: GHandler s m ()

Called on a successful login. By default, calls setMessageI NowLoggedIn.

onLogout :: GHandler s m ()

Called on logout. By default, does nothing

maybeAuthId :: GHandler s m (Maybe (AuthId m))

Retrieves user credentials, if user is authenticated.

By default, this calls defaultMaybeAuthId to get the user ID from the session. This can be overridden to allow authentication via other means, such as checking for a special token in a request header. This is especially useful for creating an API to be accessed via some means other than a browser.

Since 1.1.2

Plugin interface

data Creds m

User credentials

Constructors

Creds

Fields

credsPlugin :: Text

How the user was authenticated

credsIdent :: Text

Identifier. Exact meaning depends on plugin.

credsExtra :: [(Text, Text)]

setCreds :: YesodAuth m => Bool -> Creds m -> GHandler s m ()

FIXME: won't show up till redirect

User functions

defaultMaybeAuthId :: YesodAuth m => GHandler s m (Maybe (AuthId m))

Retrieves user credentials from the session, if user is authenticated.

Since 1.1.2

maybeAuth :: (YesodAuth m, b ~ YesodPersistBackend m, b ~ PersistEntityBackend val, Key b val ~ AuthId m, PersistStore b (GHandler s m), PersistEntity val, YesodPersist m) => GHandler s m (Maybe (Entity val))

requireAuthId :: YesodAuth m => GHandler s m (AuthId m)

requireAuth :: (YesodAuth m, b ~ YesodPersistBackend m, b ~ PersistEntityBackend val, Key b val ~ AuthId m, PersistStore b (GHandler s m), PersistEntity val, YesodPersist m) => GHandler s m (Entity val)

Exception

data AuthException

Constructors

InvalidBrowserIDAssertion

InvalidFacebookResponse