Yesod.Auth.BrowserId

Documentation

authBrowserId :: YesodAuth m => AuthPlugin m

Log into browser ID with an audience value determined from the approot.

authBrowserIdAudience

Arguments

:: YesodAuth m

=> Text
audience

-> AuthPlugin m

Log into browser ID with the given audience value. Note that this must be your actual hostname, or login will fail.

createOnClick :: (Route Auth -> Route master) -> GWidget sub master Text

Generates a function to handle on-click events, and returns that function name.