Yesod.Auth.Dummy

Description

Provides a dummy authentication module that simply lets a user specify his/her identifier. This is not intended for real world use, just for testing.

Documentation

authDummy :: YesodAuth m => AuthPlugin m