Yesod.Auth.GoogleEmail

Description

Use an email address as an identifier via Google's OpenID login system.

This backend will not use the OpenID identifier at all. It only uses OpenID as a login system. By using this plugin, you are trusting Google to validate an email address, and requiring users to have a Google account. On the plus side, you get to use email addresses as the identifier, many users have existing Google accounts, the login system has been long tested (as opposed to BrowserID), and it requires no credential managing or setup (as opposed to Email).

Documentation

authGoogleEmail :: YesodAuth m => AuthPlugin m

forwardUrl :: AuthRoute