Yesod.Auth.GoogleEmail2

Description

Use an email address as an identifier via Google's login system.

Note that this is a replacement for Yesod.Auth.GoogleEmail, which depends on Google's now deprecated OpenID system. For more information, see https://developers.google.com/+/api/auth-migration.

By using this plugin, you are trusting Google to validate an email address, and requiring users to have a Google account. On the plus side, you get to use email addresses as the identifier, many users have existing Google accounts, the login system has been long tested (as opposed to BrowserID), and it requires no credential managing or setup (as opposed to Email).

In order to use this plugin:

Create an application on the Google Developer Console https://console.developers.google.com/

Create OAuth credentials. The redirect URI will be http://yourdomain/auth/page/googleemail2/complete. (If you have your authentication subsite at a different root than /auth/, please adjust accordingly.)

Enable the Google+ API.

Since: 1.3.1

Authentication handlers

authGoogleEmail

Arguments

:: YesodAuth m

=> Text
client ID

-> Text
client secret

-> AuthPlugin m

authGoogleEmailSaveToken

Arguments

:: YesodAuth m

=> Text
client ID

-> Text
client secret

-> AuthPlugin m

An alternative version which stores user access token in the session variable. Use it if you want to request user's profile from your app.

Since: 1.4.3

forwardUrl :: AuthRoute

User authentication token

data Token

An authentication token which was acquired from OAuth callback. The token gets saved into the session storage only if you use authGoogleEmailSaveToken. You can acquire saved token with getUserAccessToken.

Since: 1.4.3

Constructors

Token

Fields

accessToken :: Text

tokenType :: Text

getUserAccessToken :: MonadHandler m => m (Maybe Token)

Get user's access token from the session. Returns Nothing if it's not found (probably because the user is not logged in via GoogleEmail2 or you are not using authGoogleEmailSaveToken)

Person

getPerson :: Manager -> Token -> HandlerT site IO (Maybe Person)

Allows to fetch information about a user from Google's API. In case of parsing error returns Nothing. Will throw HttpException in case of network problems or error response code.

Since: 1.4.3

data Person

Information about the user Full description of the resource https://developers.google.com/+/api/latest/people

Since: 1.4.3

Constructors

Person

Fields

personId :: Text

personDisplayName :: Maybe Text

The name of this person, which is suitable for display

personName :: Maybe Name

personNickname :: Maybe Text

personBirthday :: Maybe Text

Birthday formatted as YYYY-MM-DD

personGender :: Maybe Gender

personProfileUri :: Maybe Text

The URI of this person's profile

personImage :: Maybe PersonImage

personAboutMe :: Maybe Text

A short biography for this person

personRelationshipStatus :: Maybe RelationshipStatus

personUris :: [PersonURI]

personOrganizations :: [Organization]

personPlacesLived :: [Place]

personTagline :: Maybe Text

The brief description of this person

personIsPlusUser :: Maybe Bool

Whether this user has signed up for Google+

personBraggingRights :: Maybe Text

The "bragging rights" line of this person

personPlusOneCount :: Maybe Int

if a Google+ page, the number of people who have +1'd this page

personCircledByCount :: Maybe Int

For followers who are visible, the number of people who have added this person or page to a circle.

personVerified :: Maybe Bool

Whether the person or Google+ Page has been verified. This is used only for pages with a higher risk of being impersonated or similar. This flag will not be present on most profiles.

personLanguage :: Maybe Text

The user's preferred language for rendering.

personEmails :: [Email]

personDomain :: Maybe Text

personOccupation :: Maybe Text

The occupation of this person

personSkills :: Maybe Text

The person's skills

data Name

Individual components of a name

Since: 1.4.3

Constructors

Name

Fields

nameFormatted :: Maybe Text

The full name of this person, including middle names, suffixes, etc

nameFamily :: Maybe Text

The family name (last name) of this person

nameGiven :: Maybe Text

The given name (first name) of this person

nameMiddle :: Maybe Text

The middle name of this person.

nameHonorificPrefix :: Maybe Text

The honorific prefixes (such as "Dr." or "Mrs.") for this person

nameHonorificSuffix :: Maybe Text

The honorific suffixes (such as "Jr.") for this person

data Gender

Gender of the person

Since: 1.4.3

Constructors

Male

Female

OtherGender

newtype PersonImage

The URI of the person's profile photo.

Since: 1.4.3

Constructors

PersonImage

Fields

imageUri :: Text

resizePersonImage :: PersonImage -> Int -> PersonImage

resizePersonImage img 30 would set query part to ?sz=30 which would resize the image under the URI. If for some reason you need to modify the query part, you should do it after resizing.

Since: 1.4.3

data RelationshipStatus

The person's relationship status.

Since: 1.4.3

Constructors

Single
Person is single

InRelationship
Person is in a relationship

Engaged
Person is engaged

Married
Person is married

Complicated
The relationship is complicated

OpenRelationship
Person is in an open relationship

Widowed
Person is widowed

DomesticPartnership
Person is in a domestic partnership

CivilUnion
Person is in a civil union

RelationshipStatus Text
Something else

data PersonURI

URIs specified in the person's profile

Since: 1.4.3

Constructors

PersonURI

Fields

uriLabel :: Maybe Text

uriValue :: Maybe Text

uriType :: Maybe PersonURIType

data PersonURIType

The type of URI

Since: 1.4.3

Constructors

OtherProfile
URI for another profile

Contributor
URI to a site for which this person is a contributor

Website
URI for this Google+ Page's primary website

OtherURI
Other URL

PersonURIType Text
Something else

data Organization

Current or past organizations with which this person is associated

Since: 1.4.3

Constructors

Organization

Fields

orgName :: Maybe Text

The person's job title or role within the organization

orgTitle :: Maybe Text

orgType :: Maybe OrganizationType

The date that the person joined this organization.

orgStartDate :: Maybe Text

The date that the person left this organization.

orgEndDate :: Maybe Text

If True, indicates this organization is the person's ^ primary one, which is typically interpreted as the current one.

orgPrimary :: Maybe Bool

data OrganizationType

The type of an organization

Since: 1.4.3

Constructors

Work

School

OrganizationType Text
Something else

data Place

A place where the person has lived or is living at the moment.

Since: 1.4.3

Constructors

Place

Fields

placeValue :: Maybe Text

A place where this person has lived. For example: "Seattle, WA", "Near Toronto".

placePrimary :: Maybe Bool

If True, this place of residence is this person's primary residence.

data Email

Person's email

Since: 1.4.3

Constructors

Email

Fields

emailValue :: Text

emailType :: EmailType

data EmailType

Type of email

Since: 1.4.3

Constructors

EmailAccount
Google account email address

EmailHome
Home email address

EmailWork
Work email adress

EmailOther
Other email address

EmailType Text
Something else

Other functions

pid :: Text

Plugin identifier. This is used to identify the plugin used for authentication. The credsPlugin will contain this value when this plugin is used for authentication. @since 1.4.17