Yesod.Auth.OpenId

Documentation

authOpenId

Arguments

:: YesodAuth m

=> IdentifierType

-> [(Text, Text)]
extension fields

-> AuthPlugin m

forwardUrl :: AuthRoute

claimedKey :: Text

The main identifier provided by the OpenID authentication plugin is the "OP-local identifier". There is also sometimes a "claimed" identifier available.

In the credsExtra field of the Creds datatype, you can lookup this key to find the claimed identifier, if available.

 let finalID = fromMaybe (credsIdent creds)
             $ lookup claimedKey (credsExtra creds)

Since 1.0.2

opLocalKey :: Text

credsIdentClaimed :: Creds m -> Text

A helper function which will get the claimed identifier, if available, falling back to the OP local identifier.

See claimedKey.

Since 1.0.2

data IdentifierType

Constructors

Claimed

OPLocal