Yesod.Auth.BrowserId

Description

Deprecated: Mozilla Persona will be shut down by the end of 2016

NOTE: Mozilla Persona will be shut down by the end of 2016, therefore this module is no longer recommended for use.

Documentation

authBrowserId :: YesodAuth m => BrowserIdSettings -> AuthPlugin m

createOnClick :: BrowserIdSettings -> (Route Auth -> Route master) -> WidgetT master IO Text

Generates a function to handle on-click events, and returns that function name.

createOnClickOverride :: BrowserIdSettings -> (Route Auth -> Route master) -> Maybe (Route master) -> WidgetT master IO Text

Generates a function to handle on-click events, and returns that function name.

def :: Default a => a

The default value for this type.

data BrowserIdSettings

A settings type for various configuration options relevant to BrowserID.

See: http://www.yesodweb.com/book/settings-types

Since 1.2.0

bisAudience :: BrowserIdSettings -> Maybe Text

BrowserID audience value. If Nothing, will be extracted based on the approot.

Default: Nothing

Since 1.2.0

bisLazyLoad :: BrowserIdSettings -> Bool

Use asynchronous Javascript loading for the BrowserID JS file.

Default: True.

Since 1.2.0

forwardUrl :: AuthRoute