Yesod.Auth.Email

Plugin

authEmail :: YesodAuthEmail m => AuthPlugin m

class (YesodAuth site, PathPiece (AuthEmailId site)) => YesodAuthEmail site where

Associated Types

type AuthEmailId site

Methods

addUnverified :: Email -> VerKey -> HandlerT site IO (AuthEmailId site)

Add a new email address to the database, but indicate that the address has not yet been verified.

Since 1.1.0

sendVerifyEmail :: Email -> VerKey -> VerUrl -> HandlerT site IO ()

Send an email to the given address to verify ownership.

Since 1.1.0

getVerifyKey :: AuthEmailId site -> HandlerT site IO (Maybe VerKey)

Get the verification key for the given email ID.

Since 1.1.0

setVerifyKey :: AuthEmailId site -> VerKey -> HandlerT site IO ()

Set the verification key for the given email ID.

Since 1.1.0

verifyAccount :: AuthEmailId site -> HandlerT site IO (Maybe (AuthId site))

Verify the email address on the given account.

Since 1.1.0

getPassword :: AuthId site -> HandlerT site IO (Maybe SaltedPass)

Get the salted password for the given account.

Since 1.1.0

setPassword :: AuthId site -> SaltedPass -> HandlerT site IO ()

Set the salted password for the given account.

Since 1.1.0

getEmailCreds :: Identifier -> HandlerT site IO (Maybe (EmailCreds site))

Get the credentials for the given Identifier, which may be either an email address or some other identification (e.g., username).

Since 1.2.0

getEmail :: AuthEmailId site -> HandlerT site IO (Maybe Email)

Get the email address for the given email ID.

Since 1.1.0

randomKey :: site -> IO Text

Generate a random alphanumeric string.

Since 1.1.0

afterPasswordRoute :: site -> Route site

Route to send user to after password has been set correctly.

Since 1.2.0

data EmailCreds site

Data stored in a database for each e-mail address.

Constructors

EmailCreds

Fields

emailCredsId :: AuthEmailId site

emailCredsAuthId :: Maybe (AuthId site)

emailCredsStatus :: VerStatus

emailCredsVerkey :: Maybe VerKey

emailCredsEmail :: Email

saltPass :: Text -> IO Text

Salt a password with a randomly generated salt.

Routes

loginR :: AuthRoute

registerR :: AuthRoute

forgotPasswordR :: AuthRoute

setpassR :: AuthRoute

isValidPass

Arguments

:: Text
cleartext password

-> SaltedPass
salted password

-> Bool

Types

type Email = Text

type VerKey = Text

type VerUrl = Text

type SaltedPass = Text

type VerStatus = Bool

type Identifier = Text

An Identifier generalizes an email address to allow users to log in with some other form of credentials (e.g., username).

Note that any of these other identifiers must not be valid email addresses.

Since 1.2.0