Yesod.Auth.Email

Plugin

authEmail :: YesodAuthEmail m => AuthPlugin m

class (YesodAuth site, PathPiece (AuthEmailId site)) => YesodAuthEmail site where

Associated Types

type AuthEmailId site

Methods

addUnverified :: Email -> VerKey -> HandlerT site IO (AuthEmailId site)

Add a new email address to the database, but indicate that the address has not yet been verified.

Since 1.1.0

sendVerifyEmail :: Email -> VerKey -> VerUrl -> HandlerT site IO ()

Send an email to the given address to verify ownership.

Since 1.1.0

getVerifyKey :: AuthEmailId site -> HandlerT site IO (Maybe VerKey)

Get the verification key for the given email ID.

Since 1.1.0

setVerifyKey :: AuthEmailId site -> VerKey -> HandlerT site IO ()

Set the verification key for the given email ID.

Since 1.1.0

verifyAccount :: AuthEmailId site -> HandlerT site IO (Maybe (AuthId site))

Verify the email address on the given account.

Since 1.1.0

getPassword :: AuthId site -> HandlerT site IO (Maybe SaltedPass)

Get the salted password for the given account.

Since 1.1.0

setPassword :: AuthId site -> SaltedPass -> HandlerT site IO ()

Set the salted password for the given account.

Since 1.1.0

getEmailCreds :: Identifier -> HandlerT site IO (Maybe (EmailCreds site))

Get the credentials for the given Identifier, which may be either an email address or some other identification (e.g., username).

Since 1.2.0

getEmail :: AuthEmailId site -> HandlerT site IO (Maybe Email)

Get the email address for the given email ID.

Since 1.1.0

randomKey :: site -> IO Text

Generate a random alphanumeric string.

Since 1.1.0

afterPasswordRoute :: site -> Route site

Route to send user to after password has been set correctly.

Since 1.2.0

needOldPassword :: AuthId site -> HandlerT site IO Bool

Does the user need to provide the current password in order to set a new password?

Default: if the user logged in via an email link do not require a password.

Since 1.2.1

checkPasswordSecurity :: AuthId site -> Text -> HandlerT site IO (Either Text ())

Check that the given plain-text password meets minimum security standards.

Default: password is at least three characters.

confirmationEmailSentResponse :: Text -> HandlerT site IO Html

Response after sending a confirmation email.

Since 1.2.2

normalizeEmailAddress :: site -> Text -> Text

Additional normalization of email addresses, besides standard canonicalization.

Default: do nothing. Note that in future versions of Yesod, the default will change to lower casing the email address. At that point, you will need to either ensure your database values are migrated to lower case, or change this default back to doing nothing.

Since 1.2.3

data EmailCreds site

Data stored in a database for each e-mail address.

Constructors

EmailCreds

Fields

emailCredsId :: AuthEmailId site

emailCredsAuthId :: Maybe (AuthId site)

emailCredsStatus :: VerStatus

emailCredsVerkey :: Maybe VerKey

emailCredsEmail :: Email

saltPass :: Text -> IO Text

Salt a password with a randomly generated salt.

Routes

loginR :: AuthRoute

registerR :: AuthRoute

forgotPasswordR :: AuthRoute

setpassR :: AuthRoute

isValidPass

Arguments

:: Text
cleartext password

-> SaltedPass
salted password

-> Bool

Types

type Email = Text

type VerKey = Text

type VerUrl = Text

type SaltedPass = Text

type VerStatus = Bool

type Identifier = Text

An Identifier generalizes an email address to allow users to log in with some other form of credentials (e.g., username).

Note that any of these other identifiers must not be valid email addresses.

Since 1.2.0

Misc

loginLinkKey :: Text

Session variable set when user logged in via a login link. See needOldPassword.

Since 1.2.1

setLoginLinkKey :: (YesodAuthEmail site, MonadHandler m, HandlerSite m ~ site) => AuthId site -> m ()

Set loginLinkKey to the current time.

Since 1.2.1